--- Back/Schemas/userScheme.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-class MembershipRequest(BaseModel):
-    userID:str
-    password:str
-    userName:str
-class LoginRequest(BaseModel):
-    userID: str
-    password: str
