+++ front/frontCode/createAccount.html
@@ -97,7 +97,6 @@
         <!-- Bootstrap core JS-->
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"></script>
         <!-- Core theme JS-->
-        <script src="js/createAccount.js"></script>
 
     </body>
 </html>
