+++ front/frontCode/js/createAccount.js
@@ -18,7 +18,7 @@ document.addEventListener("DOMContentLoaded", function () {
         }
 
         try {
-            const response = await fetch("http://localhost:8000/membership", {
+            const response = await fetch("http://localhost:8000/signup", {
                 method: "POST",
                 headers: {
                     "Content-Type": "application/json"
